# Fix iterable length handling in contrib.concurrent