--- a/tqdm/contrib/concurrent.py
+++ b/tqdm/contrib/concurrent.py
@@ -26,14 +26,9 @@ def ensure_lock(tqdm_class, lock_name=""):
         tqdm_class.set_lock(old_lock)
 
 
-def _shortest_iterable_length(iterables):
-    # Negative default for iterables that have no length (e.g. itertools.repeat)
-    iterable_lengths = [length_hint(iterable, -1) for iterable in iterables]
-    # Remove the negative values
-    iterable_lengths = filter(lambda length: length >= 0, iterable_lengths)
-    # Take the shortest of the finite iterables
-    shortest_iterable_len = min(iterable_lengths)
-    return shortest_iterable_len
+def _min_map_len(iterables):
+    """min(map(length_hint, iterables))"""
+    return min(n for it in iterables if (n := length_hint(it, -1)) >= 0)
 
 
 def _executor_map(PoolExecutor, fn, *iterables, **tqdm_kwargs):
@@ -49,7 +44,7 @@ def _executor_map(PoolExecutor, fn, *iterables, **tqdm_kwargs):
     """
     kwargs = tqdm_kwargs.copy()
     if "total" not in kwargs:
-        kwargs["total"] = _shortest_iterable_length(iterables)
+        kwargs["total"] = _min_map_len(iterables)
     tqdm_class = kwargs.pop("tqdm_class", tqdm_auto)
     max_workers = kwargs.pop("max_workers", min(32, cpu_count() + 4))
     chunksize = kwargs.pop("chunksize", 1)
@@ -110,7 +105,7 @@ def process_map(fn, *iterables, **tqdm_kwargs):
     if iterables and "chunksize" not in tqdm_kwargs:
         # default `chunksize=1` has poor performance for large iterables
         # (most time spent dispatching items to workers).
-        shortest_iterable_len = _shortest_iterable_length(iterables)
+        shortest_iterable_len = _min_map_len(iterables)
         if shortest_iterable_len > 1000:
             from warnings import warn
             warn("Iterable length %d > 1000 but `chunksize` is not set."